ajustes al cambiar el id de long a int

+++ TiendaApiDto/Controllers/EmpleadoController.cs
@@ -40,11 +40,11 @@ namespace TiendaApiDto.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<ActionResult<EmpleadoDto>> Update(long id, [FromBody] EmpleadoDto dto)
+        public async Task<ActionResult<EmpleadoDto>> Update(int id, [FromBody] EmpleadoDto dto)
         {
             if (id != dto.Id) return BadRequest("El ID no coincide");
 
-            var existente = await _repository.GetByIdAsync((int)id);
+            var existente = await _repository.GetByIdAsync(id);
             if (existente == null) return NotFound();
 
             existente.UpdateEntity(dto);
@@ -53,7 +53,7 @@ namespace TiendaApiDto.Controllers
         }
 
         [HttpDelete("{id}")]
-        public async Task<ActionResult> Delete(long id)
+        public async Task<ActionResult> Delete(int id)
         {
             var eliminado = await _repository.DeleteAsync(id);
             if (!eliminado) return NotFound();
